requirements.txt
# ─── CDSS Backend Dependencies ───
fastapi>=0.133.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.22

# ML / Deep Learning
torch>=2.10.0
torchvision>=0.25.0
timm>=1.0.25
grad-cam>=1.5.5

# Data & Image Processing
numpy>=2.4.0
pandas>=2.3.0
Pillow>=12.1.0
scipy>=1.17.0
joblib>=1.5.0
matplotlib>=3.10.0

# HTTP & Environment
httpx>=0.28.0
python-dotenv>=1.2.0

# Schema Validation (pulled in by FastAPI, pinned for clarity)
pydantic>=2.12.0

# Azure Blob Storage (download Models & Dataset at startup)
azure-storage-blob>=12.25.0
